- Fixed warnings.

+++ src/scripting/vmthunks.cpp
@@ -1336,7 +1336,7 @@ DEFINE_ACTION_FUNCTION_NATIVE(_Sector, RemoveForceField, RemoveForceField)
 
  static F3DFloor* Get3DFloor(sector_t *self, int index)
  {
- 	 if ((index < 0) || (index >= self->e->XFloor.ffloors.Size()))
+ 	 if ((index < 0) || ((unsigned int)index >= self->e->XFloor.ffloors.Size()))
  	 	return nullptr;
 	 return self->e->XFloor.ffloors[index];
  }
@@ -1361,7 +1361,7 @@ DEFINE_ACTION_FUNCTION_NATIVE(_Sector, RemoveForceField, RemoveForceField)
 
  static sector_t* GetAttached(sector_t *self, int index)
  {
- 	 if ((index < 0) || (index >= self->e->XFloor.attached.Size()))
+ 	 if ((index < 0) || ((unsigned int)index >= self->e->XFloor.attached.Size()))
  	 	return nullptr;
 	 return self->e->XFloor.attached[index];
  }
